Update for home.java
ok so home is almost done, there are a few queries i have to work on,
need to populate the table, but the basic code is all there, if you guys
need to use as a reference go for it. if you have questions ask. i will
update later with comments in the code so you can follow a bit better if
you get lost